+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Extend Victron JSON-Data-Output
+
 ## 3.0.0-alpha.48 (2024-05-03)
 
 * (theimo1221) Basic AC-Block implementation